pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "discord-py>=2.6.4",
    "flask>=3.1.2",
    "python-dotenv>=1.1.1",
    "wavelink>=3.4.1",
]
